+++ RxExample/RxExample/Examples/GitHubSignup/UsingVanillaObservables/GitHubSignupViewController1.swift
@@ -31,9 +31,9 @@ class GitHubSignupViewController1 : ViewController {
 
         let viewModel = GithubSignupViewModel1(
             input: (
-                username: usernameOutlet.rx.text.asObservable(),
-                password: passwordOutlet.rx.text.asObservable(),
-                repeatedPassword: repeatedPasswordOutlet.rx.text.asObservable(),
+                username: usernameOutlet.rx.text.orEmpty.asObservable(),
+                password: passwordOutlet.rx.text.orEmpty.asObservable(),
+                repeatedPassword: repeatedPasswordOutlet.rx.text.orEmpty.asObservable(),
                 loginTaps: signupOutlet.rx.tap.asObservable()
             ),
             dependency: (