+++ RxSwift/Event.swift
@@ -64,3 +64,23 @@ extension Event {
         return nil
     }
 }
+
+extension Event {
+    /// Maps sequence elements using transform. If error happens during the transform .error
+    /// will be returned as value
+    public func map<Result>(_ transform: (Element) throws -> Result) -> Event<Result> {
+        do {
+            switch self {
+            case let .next(element):
+                return .next(try transform(element))
+            case let .error(error):
+                return .error(error)
+            case .completed:
+                return .completed
+            }
+        }
+        catch let e {
+            return .error(e)
+        }
+    }
+}