+++ RxSwift/Observables/Implementations/Map.swift
@@ -9,15 +9,15 @@
 import Foundation
 
 class MapSink<SourceType, O : ObserverType> : Sink<O>, ObserverType {
-    typealias Selector = (SourceType) throws -> ResultType
+    typealias Transform = (SourceType) throws -> ResultType
 
     typealias ResultType = O.E
     typealias Element = SourceType
 
-    private let _selector: Selector
+    private let _transform: Transform
     
-    init(selector: @escaping Selector, observer: O, cancel: Cancelable) {
-        _selector = selector
+    init(transform: @escaping Transform, observer: O, cancel: Cancelable) {
+        _transform = transform
         super.init(observer: observer, cancel: cancel)
     }
 
@@ -25,7 +25,7 @@ class MapSink<SourceType, O : ObserverType> : Sink<O>, ObserverType {
         switch event {
         case .next(let element):
             do {
-                let mappedElement = try _selector(element)
+                let mappedElement = try _transform(element)
                 forwardOn(.next(mappedElement))
             }
             catch let e {
@@ -108,15 +108,15 @@ class MapWithIndex<SourceType, ResultType> : Producer<ResultType> {
 #endif
 
 class Map<SourceType, ResultType>: Producer<ResultType> {
-    typealias Selector = (SourceType) throws -> ResultType
+    typealias Transform = (SourceType) throws -> ResultType
 
     private let _source: Observable<SourceType>
 
-    private let _selector: Selector
+    private let _transform: Transform
 
-    init(source: Observable<SourceType>, selector: @escaping Selector) {
+    init(source: Observable<SourceType>, transform: @escaping Transform) {
         _source = source
-        _selector = selector
+        _transform = transform
 
 #if TRACE_RESOURCES
         let _ = AtomicIncrement(&_numberOfMapOperators)
@@ -124,15 +124,15 @@ class Map<SourceType, ResultType>: Producer<ResultType> {
     }
 
     override func composeMap<R>(_ selector: @escaping (ResultType) throws -> R) -> Observable<R> {
-        let originalSelector = _selector
-        return Map<SourceType, R>(source: _source, selector: { (s: SourceType) throws -> R in
+        let originalSelector = _transform
+        return Map<SourceType, R>(source: _source, transform: { (s: SourceType) throws -> R in
             let r: ResultType = try originalSelector(s)
             return try selector(r)
         })
     }
     
     override func run<O: ObserverType>(_ observer: O, cancel: Cancelable) -> (sink: Disposable, subscription: Disposable) where O.E == ResultType {
-        let sink = MapSink(selector: _selector, observer: observer, cancel: cancel)
+        let sink = MapSink(transform: _transform, observer: observer, cancel: cancel)
         let subscription = _source.subscribe(sink)
         return (sink: sink, subscription: subscription)
     }