+++ Tests/RxCocoaTests/UITextField+RxTests.swift
@@ -14,7 +14,7 @@ import XCTest
 // UITextField
 class UITextFieldTests : RxTest {
     func testTextCompletesOnDealloc() {
-        ensurePropertyDeallocated({ UITextField() }, "a") { (view: UITextField) in view.rx.text }
+        ensurePropertyDeallocated({ UITextField() }, "a", comparer: { $0 == $1 }) { (view: UITextField) in view.rx.text }
     }
 
     func testSettingTextDoesntClearMarkedText() {